@@ -12,7 +12,7 @@
     "multipart"
   ],
   "engines": {
-    "node": ">=8.9.0"
+    "node": ">=8.12.0"
   },
   "dependencies": {
     "boom": "7.x.x",
@@ -24,7 +24,7 @@
   "devDependencies": {
     "code": "5.x.x",
     "form-data": "2.1.x",
-    "lab": "15.x.x"
+    "lab": "17.x.x"
   },
   "scripts": {
     "test": "lab -a code -t 100 -L",
