@@ -416,17 +416,17 @@ internals.pipe = function (from, to) {
 };
 
 
-internals.Counter = function () {
+internals.Counter = class extends Stream.Transform {
 
-    Stream.Transform.call(this);
-    this.bytes = 0;
-};
+    constructor() {
 
-Hoek.inherits(internals.Counter, Stream.Transform);
+        super();
+        this.bytes = 0;
+    }
 
+    _transform(chunk, encoding, next) {
 
-internals.Counter.prototype._transform = function (chunk, encoding, next) {
-
-    this.bytes = this.bytes + chunk.length;
-    return next(null, chunk);
+        this.bytes = this.bytes + chunk.length;
+        return next(null, chunk);
+    }
 };
